Init createenv sub-module Createenv is dedicated to setting-up the cluster and creating the table It makes the code more readable.
