+++ p3_datawarehouse_aws_redshift/src/__init__.py
@@ -1,3 +1,3 @@
 from p3_datawarehouse_aws_redshift.src.etl import main as etl_main
-from p3_datawarehouse_aws_redshift.src.create_tables import main as create_main
-from p3_datawarehouse_aws_redshift.src.create_cluster import main as setup_main
+from p3_datawarehouse_aws_redshift.src.createenv.create_tables import main as create_main
+from p3_datawarehouse_aws_redshift.src.createenv.create_cluster import main as setup_main
