+++ p3_datawarehouse_aws_redshift/src/createenv/create_cluster.py
